削除機能　実装未完了
