+++ app/model/TodoData.scala
@@ -5,10 +5,10 @@ import play.api.data._
 
 //  新規登録用Formのケースクラスと、マッピングさせた変数
 case class TodoData(
-                     title:        String,
-                     body:         String,
-                     categoryName: String
-                   )
+     title:        String,
+     body:         String,
+     categoryName: String
+)
 
 object TodoData{
   val todoForm = Form(
