+++ app/views/error/page404.scala.html
@@ -1,0 +1,4 @@
+@(vv: model.ViewValueCommon)
+@common.Default(vv){
+    <p>予期せぬエラーが発生しました。</p>
+}
