+++ app/views/todo/UserList.scala.html
@@ -17,6 +17,7 @@ userCategory:   List[UserCategory]
                 <th>現状   </th>
                 <th>カテゴリ</th>
                 <th>編集   </th>
+                <th>削除   </th>
             </tr>
             @*   for文でuserInfoの要素数、情報を出力    *@
             @for(user <- userInfo){
@@ -26,7 +27,8 @@ userCategory:   List[UserCategory]
                     <td>             @user.state.name </td>
                     @*   背景色変換とカテゴリ名の出力は、userList.jsで操作     *@
                     <td class="slug">@user.category_id.toString</td>
-                    <td><a href="@controllers.routes.HomeController.edit(user.id.get.toInt)">編集画面へ</a></td>
+                    <td><a href="@controllers.routes.TodoController.edit(user.id.get.toInt)">このTodoを編集</a></td>
+                    <td><input type="submit" value="削除"></td>
                 </tr>
             }
         </table>
