+++ public/stylesheets/error/error.css
@@ -1,0 +1,43 @@
+body {
+  margin:   0;
+  padding:  0;
+}
+
+.header {
+  height:            70px;
+  background-color:  #24292F;
+  width:             100%;
+}
+
+.header .header__content {
+  width:      1080px;
+  margin:     auto;
+}
+
+.header .header__logo {
+  color:             white;
+  line-height:       70px;
+  font-size:         20px;
+  font-weight:       bold;
+}
+
+.main {
+  width:            980px;
+  margin:           100px auto;
+  padding:          100px;
+  box-shadow:       0 0 4px gray;
+  position:         relative;
+  background-color: black;
+  color:            red;
+}
+
+.success-message {
+  margin:      30px 0;
+  text-align:  center;
+}
+
+
+/******* フォームのinfoを非表示に *******/
+.info {
+  display: none;
+}
