+++ public/stylesheets/todo/userList.css
@@ -76,3 +76,23 @@ h1 {
     margin:     0;
     padding:    60px
 }
+
+.userList td:last-child {
+    padding:            0;
+    background-color:   black;
+    text-align:         center;
+}
+
+.userList td input {
+    color:              red;
+    background-color:   black;
+    border:             none;
+    height:             160px;
+    width:              100px;
+    font-size:          15px;
+    padding:            0;
+
+}
+
+
+
